C++ starter: fix bugs
The current C++ starter AI does not work well because of a bug in
`Tree::input`. The member function reads input values from `std::cin`,
but it does not stores the values to the object at all. This commit
fixes the bug.

Besides, this commit deletes the constructor
`Cell::Cell(int, int, int)`, which is a little confusing, because the
third argument represents just the size of the neighbors. This commit
also updates some other things as an improvement.